requirements.txt
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-groq
faiss-cpu
chromadb
pypdf
python-dotenv
streamlit
tiktoken